Fix: Remove extra is new login epilogue screen enabled for tracking
This is actually a fix since the tracking should not happen if the
'mLoginEpilogueListener' is null, which actually should never be the
case and might result to a bug. As such, when such case arise (if ever)
and given the fact that the main action(s) will not be performed
('onCreateNewSite()' or 'onContinue()'), when that happens the tracking
shouldn't be performed as well.

PS: The previous version of 'setupBottomButton(...)', the one before
'Onboarding Improvements' were introduced to the 'Login Epilogue' screen
was also having the very same problem.