Feature: Add show and return to site picker adapter and position logic
Although this is not necessary at the moment, since the footer
visibility will be gone anyway and as such this check redundant, if at
any point in the future this 'showAndReturn' login changes on the footer
then a bug might arise where the last element of the recycler view might
end up being not clickable as it will become an invalid position.

This change guards from the above and make the whole 'showAndReturn'
implementation on the alternative flow more robust, that is when the
'Onboarding Improvements' are being utilised.