Test: Fix e2e login tests by updating the login flow for confirm login
Since the 'Onboarding Improvements' feature flag has been turned on and
merged into 'develop', these 'LoginTests' have been started to fail.
This is expected since all the login tests were not being built to
interact with the new 'Login Epilogue' screen.

This commit updates the 'Login Flow' and more specifically the
'confirmLogin(...)' functionality so that it respects the new 'Login
Epilogue' screen. As such:
- For the WordPress app and a wp.com site, the new 'Login Epilogue'
and the 'Quick Start Prompt' are being asserted.
- For the WordPress app and a self-hosted site, the new 'Login Epilogue'
is being asserted but the 'Quick Start Prompt' is excluded from it.
- For the Jetpack app, the existing 'Login Epilogue' is asserted.