Feature: Show old and new quick start dialog on improved my site
This logic was copy-pasted as is from within the
'WPMainActivity.showQuickStartDialog(...)' method.

The initial 'if/return' logic was being migration to
'MySiteViewModel.showQuickStartDialog(...)' and was updated to become
more readable and testable, but other than that wasn't changed
logically-wise.

As such, this feature now works exactly as it used to from within the
'MySiteFragment'.