Feature: Add quick start prompt click interface on improved my site
This is done so that the 'onPositiveClicked' 'onNegativeClicked' and
'onNeutralClicked' calls get to be connected with the
'ImprovedMySiteFragment', as they were used to be connected to the
'MySiteFragment'.

Since, the 'QuickStartPromptDialogFragment' works using this kind of
pattern, the first step is to migrate it as is, become refactoring this
dialog to work with the new pattern, the one the
'ImprovedMySiteFragment' utilizes (see 'BasicDialogViewModel').