Feature: Redirect login epilogue show quick start to improved my site
Since this is the pattern that is used for handling activity results on
the 'ImprovedMySiteFragment', this was the one piece that was missing
from the redirection. Now, everything 'MySiteFragment' activity
result related is being handled on the 'MPMainActivity' layer, while
everything 'ImprovedMySiteFragment' acitivity result related is being
handled on the 'ImprovedMySiteFragment' itself, which actually makes is
more manageable in the end.