Refactor: Rename start quick start to include the check prefix
This is done due to the fact that the actual 'startQuickStart()' needs
to be reserved for later on. Also, it makes sense to add the 'checkAnd'
suffix due to the fact that this function is checking whether the
'Quick Start Dynamic Cards' feature flag is on before starting quick
start.