Refactor: Split wp main activity set selected site into set and remove
The 'setSelectedSite(...)' method is actually doing a remove as well,
and that is done when the site passed to this method is null. As such,
by splitting this method into two, a 'setSelectedSite(...)' and a
'removeSelectedSite()', this functionality is made explicit and in
addition the nullable site parameter now become non-null, which makes
the whole method more robust and tightens the api to remove extra space
for mistakes.

In addition to this change, this robustness and tightening of the api
was moved on the 'updateSite(...)' function within the repository
itself. There, the 'updateSite(...)' was split into two as well, a
'updateSite(...)' and a 'removeSite()', following by making the site
parameter non-null on the 'updateSite(...)' function as well.