Refactor: Update key local id constant to key site local id
This is done in order to make it explicity that this 'KEY_LOCAL_ID' is
actually referring to a 'SITE' local id and not some random local id.
Also, this change will utlimately make the use of 'siteLocalId' vs. the
incorrect use of 'siteId' more consistent.

All 'id', incorrect 'sideId' that are actually an 'id', 'localId' and
any variation of those will be renamed to 'siteLocalId' for consitency
purposes.