Refactor: Replace selected site storage with selected site repository
The 'SelectedSiteStorage' is only used to hold the 'currentLocalSiteId'
field, which it gets from the prefs. However, this is not necessary as
'SelectedSiteRepository' can be utilized for the exact same reason. As
such the 'SelectedSiteStorage' is completely removed since it is not
used anywhere else but within the 'StatsSiteProvider' and for that
specific reason.