Update diff calculation in keyboard resize view util
Diff is used for adding padding to the pre-publishing bottom sheet content view when the keyboard is visible so that the publish visible is not hidden behind the keyboard. It is approximately equal to the keyboard height.

Based on the runtime values, it seems windowVisibleDisplayFrame height includes the top inset when the screen is in full screen immersive mode and subtracting it from the display height will give a shorter keyboard height. Hence it is subtracted from the real screen size, and also the bottom inset is subtracted to get the actual keyboard height.