Refactor: Move update of app prefs select site to repository
Also, as part of this move, a JavaDoc was added on top of the
'AppPrefs.setSelectedSite()' to make sure that future usages of this
method are being avoided and going through the
'SelectedSiteRepository.getSelectedSite()' instead.

In addition, as part of this update, two possible bugs where fixed due
to the fact that the 'WPMainActivity.onSiteChanged(...)' and the
'WPMainActivity.refreshCurrentSelectedSiteAfterEditorChanges(...)'
methods were not updating the 'AppPrefs.setSelectedSite(...)' when that
was happening on the repository level, through
'SelectedSiteRepository.updateSite(...)'. This means, that at that
point, the state of selected site was different, the persisted and the
in-memory selected site was different. This update fix this probably bug
that might have been causing unknown troubles. However, there is a
change that this is not a bug, but an actual intended feature. Thus,
this change needs to be thoughoutly tested to verify correctness.