Feature: Increase quick start notice default duration from default 3.5 secs to twice of it = 7 secs
The deprecated MySiteFragment was displayed for 10 sec with an explicit negative button which gave it a undismissable dialog like look. In the generic snackbar, negative button is not shown, users can still dismiss it using the swipe-to-dismiss action. Chosen a time between default 3.5 sec and the 10 sec so that snackbar is not shown for too long and the duration is not reduced too much which can impact the quick start success rate.