Update public release lanes to not require app:
We don't want to have to specify app: when calling those lanes in the CLI anymore.

Some of the release-toolkit actions still require an app: parameter even if they don't use it – because we are still WIP on cleaning things up there – so for those we just pass a dummy empty value instead.
Which will also allow us to check during the next run that this value indeed doesn't matter, before removing it completely