Remove legacy, unused `trigger_ci` private lane
This has been replaced by the public `trigger_release_build` lane, which makes use of the `circleci_trigger_job` action implemented in toolkit.

The old `trigger_ci` lane was not used anywhere in the Fastfile anymore (and being a private lane, couldn't be invoked manually from CLI either), confirming it was not used anymore.