Fixes: Retry logic in CategoriesListViewModel
Fixes retry logic in CategoriesListViewModel by moving the UI state update of loading to outside network check
Adds a retry delay so that a loading is displayed when internet is not available and retry is invoked. If internet is not available and retry is invoked, It will look like the Retry button is not working. So a retry delay shows loading indicator for a duration.
Updates the invoking function on retry so that a delay can be added in the coroutine.