Refactor: Update cards model structure
The current 'CardsModel' structure doesn't fit the agreed database
structure, where every card type will be inserted in its own row. As
such, the structure is now changing to the new 'CardModel' sealed class
structure, where every inner data class to become the a card model with
a specific type, ultimately converting the 'cards' from a single data
class, that holds in it all the individual cards, to a list of cards.