Feature: Update post response and model
After a discussion, it was decided that the post response needs to be
updated with the newly agreed structure (a more compact post response).

You will also notice that the 'title' and 'content' are now not-null.
When there exist no title or content for a post the post response will
return an empty string (instead of a null). This effectively make
'featured_image' the only nullable property for a post card.

PS: The API endpoint has been already updated to adhere with the new
requirements, which is to return a more compact list of posts, for more
details see D69742.