Refactor: Rename post cards source to cards source
This source is actually going to be used for all dashboard related
cards, be it 'posts', 'stats', 'pages', 'comments', etc. As such, the
more specific 'post' naming is no longer appropriate and can be omitted.

Also, as with all the other 'cards' related naming convention, the
'Dashboard' word is omitted as well and used on the package level. Thus,
this makes the naming of the source as simple as possible, in this case
just 'CardsSource'. This also fits well with the 'CardsStore' on the
FluxC side.