Refactor: Replace mocked posts data and json utils with cards result
Instead this commit replaces the 'MockedPostsData' with the actual
'CardsResult<List<CardModel>>' model, which the cards source is to
return.

At this moment, the 'CardsSource' return an empty 'CardsResult', but
subsequent commits will build on top of that to return either a non
empty or an error 'CardResult'.

Also, this commit updates the 'PostCardBuilder', which now uses the
actual data to build the individual post cards.