Feature: Implement cards source logic based on get data and error
The way 'CardsSource' will provide the 'CardsUpdate' will be twofold:
1. First, the on 'build(...)' will get triggered. At that point the
'CardsStore' will hook on getting the cards from the database,
effectively collecting them from Room as a flow. At this point, any
available persisted/cached data will be displayed on screen immediately.
2. Then, and right after, a refresh of data will get triggered,
which will happen either immediately, through the initial 'build(...)'
trigger, or when 'refresh' get triggered on demand (for example due to
the user initiating a pull-to-refresh, or when the screen gets resumed).
At this point the 'CardsStore' will trigger a network request to fetch
the cards from the backend. When the network request returns, the
updated cards will be persisted in the database and the 'getData(...)'
hook will get automatically triggered (see 1. above).

In case of an error, during the fetching of the cards from the network,
the error state will be posted as well. Else, the success and empty
state of that network request will be ignored as it will be anyway
consumed by the flow, through Room (see 1. above).

This commit makes the previous failing test pass (TDD).