Feature: Add cards gson date adapter and convert date to gmt timezone
The API endpoint date for the post cards are in GMT format and as such
this enhancement makes sure that this is being respected.

As such:
- First, when converting the 'date' post response string using the
  'fromDate(...)' function, the date format timezone is set to 'GMT'.
- Then, when converting the card entity from and to a JSON using GSON in
  order to save and retrieve the card model from the database, a newly
  created, private to cards 'GsonDateAdapter' is used which also sets
  the date format timezone to 'GMT'.

Both additions, makes sure that the date ends up being the expected one
when it is being utilized later on and converted to the user's timezone.