Feature: Add navigate to post functionality to post list view model
This functionality is a copy-paste of the existing 'scrollToPosition'
functionality, which also happens when the 'onDataUpdated(...)' get
triggered.

The difference here is that now the 'RemotePostId' is used to find the
item in the list, instead of the 'LocalPostId'. Also, another difference
is that since the 'onDataUpdated(...)' can get triggered multiple times,
the 'navigateToRemotePostId' is only reset when the item is found. This
can happen either when 'onDataUpdated(...)' is triggered for the first
time (database) or afterwards in a subsequent trigger (network).