Hack: Add extra navigate to edit post delay to not pick a stale item
In addition to the 'PostListViewModel.navigateToPost(...) and its logic,
which includes:
- The 'PostListViewModel.navigateToEditPost' to skip the first result,
  and
- The [PostListViewModel.navigateToRemotePostId] to reset the navigation
  trigger.
An extra delay of 'NAVIGATE_TO_EDIT_POST_DELAY_MS' milliseconds it added
to make sure that the list is submitted to the [PostListAdapter] paging
related adapter (see [PagedListAdapter]). This will make sure that the
newly submitted list updated the existing and potentially stale list.

Otherwise, and due to the fact that the list submission can take some
additional time, selecting the item might launch the 'EditPostActivity'
with a stale item and thus this might overwrite any remote changes.

PS: A delay of half a second would have been sufficient enough, based on
the testing that was done, but doubling this number to 1 second seemed
safer to to make sure that no there will be less changes for the stale
item to get selected.

NOTE: This whole solution is hacky and done this way to help us achieve
a quick result, without needing to update the paging library or the
underneath list store related implementation.