Set refresh initial value to true
This is done because a site is selected the first time on WPMAinActivity's resume whereas my site sources are created before that. Based on the context we have, rather than selecting the site in the WPMainActivity's onCreate (which is done in a few scenarios like opening from shortcut), it feels more safe to update my site sources' refresh initial value to true.

This means the initial loading indicator will get visible again for no site scenario. This is fixed as part of the next commit.