Feature: Navigate to draft/scheduled posts on post item click
This navigation is temporary and as such not utilizing the
'action.postId' in order to navigate to the 'Edit Post' screen. Instead,
it fallbacks to navigating to the 'Posts' screen and targeting a
specific tab.

Previously, another issue was opened in order to help us create a
solution to navigate straight to 'Edit Post', just as the user expected.
However, as this solution was implemented and problems were popping-up,
this solution was paused indefinitely and until it further reassessment.

Useful links:
- Issue: https://github.com/wordpress-mobile/WordPress-Android/issues
/15709
- PR: https://github.com/wordpress-mobile/WordPress-Android/pull/15720
- PR: https://github.com/wordpress-mobile/WordPress-Android/pull/15723
- Decision: https://github.com/wordpress-mobile/WordPress-Android/pull/
15723#issuecomment-996762177