Override ZendeskRequestListButton button style
ZendeskRequestListButton style is appled to "Start a conversation" button on the Zendesk `My Tickets` (empty) screen. The style is overriden to fix the invisible button issue.

It seems that the `...NoActionBar` theme does not allow `backgroundTint` to be applied properly. We should either change the Zendesk activity theme to a `...NoActionBar.Bridge` theme or use a workaround to set the backgroundTint to null and set a custom background.

Reference:
https://git.io/JyEnY
https://stackoverflow.com/a/67862797/193545