Feature: Implement authorization required case on cards store
With this change, when a user gets downgradeted from 'administrator',
'editor', 'author' or 'contributor' to a 'follower' or 'subscriber',
then it is being made sure that on an authorization required error:
1) The database is being clean of any cards model from previous
   requests, and
2) Instead of having the user seeing an error message, no error message
is shown. Instead the dashboard cards feed just disappears from the
   'My Site' tab.

This commit makes the previous failing test pass (TDD).