Revert "Update strings.xml"
This reverts commit 4c3b085742eec1aa8509bfd2f0aa5bed030af2bc.

The strings from libraries should not be removed or edited from the app's `strings.xml`. Those are needed to be in the app's own `strings.xml` because of our translation pipeline requires it. They are copied over from the libs' `strings.xml` into the app's one during code freeze and should remain there in order to be translated as expected. This is a glitch of our translation pipeline that we hope to address at some point (see pxLjZ-6Fc-p2).