Analysis: Suppress notification trampoline for post upload notifier
This is currently suppressed because:
- Currently, this isn't causing any crashes or the app misbehaving.
This issue can just lead to significant lags in some scenarios.
- Also, it is not trivial to fix this and thus out of scope of the
Gradle and AGP upgrade. The activity should be directly launched from
the notification.

However, this issue will need to be fixed sooner than later, especially
because of the fact that it will become forbidden when the
'targetSdkVersion' will be set to API level for Android 12 or higher.
Thus, a GitHub issue was created to accompany this commit and issue
suppression. From more info see the newly created GitHub issue below:

https://github.com/wordpress-mobile/WordPress-Android/issues/15875