Analysis: Suppress lint error for html to spanned converter
The lint detector androidx.appcompat
.BaseMethodDeprecationDetector$ApiAbove called context.getMainProject()
during module analysis.

This does not work correctly when running in AGP (7.0.4).

In particular, there may be false positives or false negatives because
the lint check may be using the minSdkVersion or manifest information
from the library instead of any consuming app module.

Contact the vendor of the lint issue to get it fixed/updated (if
known, listed below), and in the meantime you can try to work around
this by disabling the following issues:

Call stack: Context
 .getMainProject(Context.kt:117)
←BaseMethodDeprecationDetector$ApiAbove
 .matches(BaseMethodDeprecationDetector.kt:58)
←BaseMethodDeprecationDetector$DeprecationCondition
 .matches(BaseMethodDeprecationDetector.kt:136)
←BaseMethodDeprecationDetector
 .visitMethodCall(BaseMethodDeprecationDetector.kt:158)
←UElementVisitor$DelegatingPsiVisitor
 .visitMethodCallExpression(UElementVisitor.kt:1128)
←UElementVisitor$DelegatingPsiVisitor
 .visitCallExpression(UElementVisitor.kt:1108)
←UCallExpression$DefaultImpls
 .accept(UCallExpression.kt:85)
←UCallExpressionEx$DefaultImpls
 .accept(UCallExpression.kt:-1)
←JavaUCallExpression
 .accept(javaUCallExpressions.kt:26)
←UQualifiedReferenceExpression$DefaultImpls
 .accept(UQualifiedReferenceExpression.kt:34)
←JavaUCompositeQualifiedExpression
 .accept(JavaUCompositeQualifiedExpression.kt:23)
←UVariableKt
 .visitContents(UVariable.kt:64)
←UVariableKt
 .access$visitContents(UVariable.kt:1)
←ULocalVariable$DefaultImpls
 .accept(UVariable.kt:110)
←ULocalVariableEx$DefaultImpls
 .accept(UVariable.kt:-1)
←JavaULocalVariable
 .accept(JavaUVariable.kt:92)
←ImplementationUtilsKt
 .acceptList(implementationUtils.kt:23)
←UDeclarationsExpression$DefaultImpls
 .accept(UDeclarationsExpression.kt:22)
←JavaUDeclarationsExpression
 .accept(JavaUDeclarationsExpression.kt:7)
←ImplementationUtilsKt
 .acceptList(implementationUtils.kt:23)

This issue type represents a problem running lint itself. Examples
include failure to find bytecode for source files (which means certain
detectors could not be run), parsing errors in lint configuration
files, etc.

These errors are not errors in your own code, but they are shown to make
it clear that some checks were not completed.