Fix for stats endpoints being called twice
The previous fix #15941 fixed most of it.  There were some duplicate calls happening when you switch to Day tab from Inisghts tab first time.

It is the same animation issue that was fixed on StatsListFragment.  However, StatsListFragment layout is being used as an include on StatsViewAllFragment and it seems to cause the same issue here.