Feature: Access tab fragment and pass quick start prompt click actions, activity result to it
Note: There was no clean way to get current tab fragment from ViewPager2 and it's adapter, an instance of FragmentStateAdapter without having to manage fragment's state.
See:
1. Feature Request: https://issuetracker.google.com/issues/202423436
2. SO Thread: https://stackoverflow.com/questions/55728719/get-current-fragment-with-viewpager2

Since FragmentStateAdapter manages the saving and restoring of the fragment's state, the tab fragment is retrieved by the tag by which the fragment is added by the adapter.