Kotlin: Remove unnecessary experimental stdlib api annotation
This '@ExperimentalStdlibApi' annotation was added all over the place
due to the 'StringsJVM.capitalize' and 'StringsJVM.decapitalize' utility
methods, which are actually no longer experimental. Thus, this
annotation is no longer necessary and can be removed from everywhere.

This is removed because otherwise the build fails to compile with the
below error:

'Opt-in requirement marker annotation on override requires the same
marker on base declaration'