Init selected tab listener only once instead of everytime it is accessed
When selectedTabListener is accessed using StatsFragment->selectedTabListener get() accessor, earlier a new instance of it got created. This newly created tab selected listener was not present the tab layout listeners and attempt to remove it inside StatsFragment->handleSelectedSection would not remove the already added listener. Multiple listeners caused duplicate events when StatsFragment->statsPager.setCurrentItem was invoked as the previous listener was not removed.

This commit fixes it by initializing the selected tab listener only once.