Track initial section selection after fix in previous commit
Previously, StatsFragment->statsPager.setCurrentItem invoked onSectionSelected (as the previous tabs selected listener was not removed) that tracked initial section selection. With the fix in previous commit, this tracking got lost and is re-added with this commit.

Note that initial section is the default Stats section to be shown when Stats screen is accessed. It is mapped from the ARG_DESIRED_TIMEFRAME in the Intent. It is usually set when the default Stats section to be shown is not the default INSIGHTS section present at 0th index of the tabs. It might be null otherwise. This change still does not track default INSIGHTS section selection when default initial section value is null or ARG_DESIRED_TIMEFRAME is not set in the Intent, keeping the existing behavior.