Update copies used in screenshot
In order to start translating the new copies used in pbArwn-46A-p2 in anticipation

Note that the `fastlane/screenshots.json` file — which uses those copies once translated to do the image composition — will be a bit out of sync given these updates, still referencing the old background images while the txt files will now have new content.

This is all right as we plan to update `screenshots.json` with the new images and frames for the new screenshots once we are ready to get those localized. But I still wanted to send the new copies to translation sooner rather than later so that we get the copies for all locales by the time we work on updating `screenshots.json` and the corresponding UITests generating the raw screenshots, hence the update of the `.txt` and `.po` file ahead of time.