TODO to remove `android_create_xml_release_notes`
Now that this is automated via fastlane supply (aka `upload_to_play_store` during the `upload_build_to_play_store` lane, we should not be needing this anymore.

Keeping it for a couple of sprints still just in case as a fallback, until we validate with a couple of releases that the new way works without any issue, but we should delete that in a release or two once the new way has made its proofs.