Build: (GE) Set a root project name
Context: The 'rootProject.name' effectively assigns a name to the build
as a whole, which is used in reports like build scans. If the root
project name is not set, the name will be the container directory name,
which can be unstable (i.e. you can check out your project to any
directory).