Build: (GE) Make configuration work when using a self-signed certificate
Context: If your Gradle Enterprise server uses an SSL certificate that
is not trusted by your build’s Java runtime, you will not be able to
publish build scans without explicitly allowing untrusted servers.

Use of this configuration is a security risk as it makes it easier for
a third party to intercept your build scan data. It should only be used
as a short term workaround until the server can be configured with a
trusted certificate.