Build: (GE) Apply the custom user data gradle plugin
This is done in order to capture more data required for the trial.

Context: The Common Custom User Data Gradle plugin for Gradle Enterprise
enhances published build scans by adding a set of tags, links and custom
values that have proven to be useful for many projects building with
Gradle Enterprise.