Explicitly set nulls when skipping
This is important in case the user backtracks and then decides to skip,
since the final site creation data should reflect that intention.
Without this, we'd still have the stale data at the end of the flow.
While only using this for tracking so far, we have not encountered this,
but this staleness would be an issue in the future.