Manage keyboard and focus between screens
This auto-focuses the input, since it is the only input on the screen.
It also manages the visibility of the keyboard, and restores the input
text contents from the view model (important when a user back tracks to
a previous step in the flow and then wants to continue forward).