Replace url span with SiteNavigationAction pattern to invoke onGetMoreViewsClick function
Removed URLSpan in VH for the card otherwise ClickableSpan.onClick wouldn't be triggered and corresponding action in the VM wouldn't be called which is needed for tracking.

URLSpan is still useful to get start and end index of the text to be linked. Once we get them, it can be safely removed.

After URLSpan is removed, navigation is now handled using SiteNavigationAction in the VM.