Add TODOS for loose ends from removed UI
Since we are removing some parts of an implementation that shares a base
class with the page picker, there will be some changes needed to adapt
these for the new flow. The intent with these comments is to serve as a
place-holder denoting some things that need to be rewired.