Add bottom helper text to theme picker
This implements an optional footer parameter to layouts recycle view,
useful for working around scroll issues in the screen where it is used.
It introduces a new abstraction `LayoutRowViewHolder` from which
`LayoutItemViewHolder` and `LayoutFooterViewHolder` extend.

The footer is inflated when needed from the passed resource id (the
overriden item count is incremented when not null).

It also adds the footer as a `LayoutCategoryViewType` so that it isn't
mixed with the regular items during view recycling.