Add temporary feature flag to facilitate testing
In order to make testing more convenient, this adds a temporary flag
that can be toggled in the debug settings of the app. It also adds a
hard-coded behavior to clear the Glide cache (memory and disk), for the
purpose of testing the flow with a fresh start each time.