Disable volley cache for MShots loader
For some reason, deleting the Glide cache was not enough, and disabling
the volley cache was still not enough to prevent a cache hit. I suspect
that there is an additional cache later on the network side of things.
Nevertheless, we may want to keep the volley cache disabled, to prevent
3 layers of duplicate cache.