Add reminder note about nested recycler viewholder pool
The nested recycler views can only utilized a shared pool if they are
given the same instance of the pool. We can either remove the line, or
keep a reference to the instance at a higher level in the hierarchy, so
that the children recycler views can share a pool of view holders.