Analysis: Resolve null safe mutable live data for show error message
The 'showErrorMessage' mutable live data value is set to
'event.error.message' on multiple occasions. However, 'event.error' can
be nullable and as such the 'event.error.message' itself.

There is an 'if (event.isError) { ... }' statement above to guard from
that ever happening, but the compiler is not smart enough to understand
that. Maybe with 'Kotlin Contracts' this problem would have been
avoided, but this feature is a separate discussion of its own.

Thus, by guarding this with 'let' before setting this 'showErrorMessage'
mutable live data is the best that can be done at this point of time.