Analysis: Resolve null safe mutable live data for feature announcement
The 'currentFeatureAnnouncement' mutable live data can be set to null
during 'loadFeatures()' and via
'FeatureAnnouncementProvider.getLatestFeatureAnnouncement(...)' which
might produce a nullable 'FeatureAnnouncement'. As such, the
'currentFeatureAnnouncement' itself needs to be declared as nullable.

The rest of the changes are propagated due to the above change and are
mainly guarding the rest of the functionality that depends on that with
'let' and '?' scope functions.