Refactor: Remove unused live data extensions (from 2017)
This 'LiveDataExtensions' is not used anywhere within the codebase and
as such can be safely remove. If at any point in the future it will be
needed again, engineers should either bring it back, with a more model
version of it, or use another utility from the 'LiveData' library
itself, which will serve the same purpose.

This also solves two 'NullSafeMutableLiveData' lint errors related to
this function.