Analysis: Resolve null safe mutable live data for photo picker
The 'selectedIds' mutable live data can be set to null during the
'start(...)' and via its 'selectedIds' parameter which is nullable. As
such, the 'selectedIds' itself needs to be declared as nullable.

Although the initial logic was different and what I am describing above,
and setting the 'selectedIds' was guarded with 'let' I chose to update
this logic and align that with the 'MediaPickerViewModel' logic and the
update I did there with my previous commit. Actually, this makes sense
as this view model and its associated fragment is treating this
'selectedIds' live data value as nullable already and guarding against
it accordingly.