Analysis: Resolve null safe mutable live data for site creation
The 'onIntentSelected' mutable live data can be set to null during
'onCustomVerticalSelected()' because of the 'IntentsUiState.searchQuery'
as it is actually defined as nullable. As such, the 'onIntentSelected'
itself needs to be declared as nullable.

However, this view model's associated activity, main view model and
interface were expecting a non-nullable intent. But, this can actually
change easily and the interface intent can be defined and used as
nullable as well. It ends us anyway being using within the
'SiteCreationMainVM.onSiteIntentSelected(...)', which can accept a
nullable intent without any problems.