Analysis: Suppress null safe mutable live data for stats view all
Suppress this 'NullSafeMutableLiveData' instead of changing this
'LiveData.observeEvent(...)' extension function and as such its
signature seems to be the safest choice at the moment since this
extension function is being utilized in extension places.

For more info see 'StatsViewAllFragment' and the usage of
'viewModel.showSnackbarMessage.observeEvent(...)' which prevents us from
changing this easily.