Analysis: Suppress false positive null safe mutable live data errors
For these classes in this commit the 'NullSafeMutableLiveData' Lint
error is suppressed are false positive. It is false positive because the
nullability of the variable is already checked before it is assigned to
the mutable live data field, but it is anyway ignored as being
potentially null and thus reported.