Build: Remove duplicate lifecycle declaration and update to 2.4.1
As part of this update the below occurred:
- The 'lifecycle-extensions' artifact and APIs got removed as it has
been deprecated since '2.2.0' and it is no longer updated. The
recommendation nowadays is to instead add dependencies for the specific
lifecycle artifacts needed.
- Update 'ViewModelProvider.Factory.create(...)' signature to a non-null
'ViewModel' generic as it is now defined as such by the Java
factory interface.

The above were the only breaking change. Apart from that everything else
seems to be compiling and working as expected (did a quick smoke test).
However, this update added 10 new 'NullSafeMutableLiveData' lint errors
that will be fixed in subsequent commits.