Return with no action if reply text is empty
Opening comments screen when there's an empty message looks like a fail-safe. We cannot reply with an empty message either from the app or from the website as send button is disabled till we enter a text. Even in the case of voice reply from a wearable, an empty voice reply doesn't make sense.

User can tap the notification itself to go straight to the comments screen in case there's ever an empty reply and so this logic to open comments activity from service on empty reply is removed.