Build: Remove unnecessary release build type config from editor example
The 'proguard-rules.pro' is removed as well. This file is unnecessary as
it actually has no rules enabled anyway.

When used, this 'editor-example' app module would be probably being
utilized with the 'debug' build type. And, even if it would be utilized
with the 'release' build type, just to make sure the release version
builds and works as expected, this basic release related configuration
that is now removed wouldn't be required at that point.