Build: Remove unnecessary editor example module level
Everything within the 'editor/example' module ('build.gradle' +
'src') is now moved to the new root 'editor-example' module.

With this change the new 'editor-example' module becomes less complex as
the structure is now flattened.

Also, this module now becomes active and it is now part of the root
level 'settings.gradle' includes related configuration. However, be
mindful that assembling 'editor-example' is not working at the moment as
 the build configuration is outdated and thus needs to be first updated.

For example, running the below command will produce an error on this
module's 'build.gradle' file at line 32 as the path to  the
'WordPressEditor' project needs to be first updated. Then, some missing
repositories need to be added to the configuration as well. And finally,
the codebase itself needs to be updated as there are compilation errors
too. As such, it is recommended to exclude this work for now and
progress with it on as a separate task afterwards.

Command: ./gradlew libs:editor-example:assembleDebug