Build: Add missing repositories configuration to editor example module
As per my previous commit message, the 'editor-example' app module is
not able to build at the moment. This change adds the missing
repositories that need to be added to the configuration in order to make
it build. Also, you will now notice that this repositories configuration
is matching to that of the 'editor' lib module.

However, be aware that the Java code itself needs to be updated as there
are still some compilation errors (21 in total). Unless those are fixed,
the below command will continue to fail.

Command: ./gradlew libs:editor-example:assembleDebug