Scale screenshot dimensions down to allow endpoint to scale them up
We use the `ThumbDimensionProvider` for displaying as well as fetching
preview screenshots with the correct dimensions. The values are
automatically scaled based on screen density on the client side, but for
legacy reasons, the API endpoint *also* scales the dimensions, resulting
in the scale being applied twice. To avoid this, we divide by the scale
factor when computing the dimension parameters to fetch the designs.

This can be revisited in the future if we change or version the REST
endpoint to remove the middleware scaling behavior.