Fix: Remove only the default initializer to provide custom configuration
This change was necessary because of the 'androidx.lifecycle' update to
'2.4.1' and the fact that with it the 'lifecycle-process' now uses
'androidx.startup' to initialize the 'ProcessLifecycleOwner'.

As such removing all initializers are no longer an option, as it will
also remove the 'ProcessLifecycleInitializer' initializer. Instead
removing only the default 'WorkManagerInitializer', while keeping the
rest with the 'tools:node="merge"' option, is recommended for when using
a custom configuration.

For more info visit:
- https://developer.android.com/topic/libraries/architecture/
workmanager/advanced/custom-configuration
- https://developer.android.com/jetpack/androidx/releases/
lifecycle#2.4.1