Build: Update compile sdk to 31 for analytics module
PS: No changes were needed to make this update possible.

This is the default and root level 'compileSdkVersion' that all modules
should be pointing to from now.

This is done in order to make sure that there is no discrepancy in
relation to compile SDK versions being different for different modules.

Also, and just as an FYI, for 'minSdkVersion' and 'targetSdkVersion'
this is already the case.