Protect against wrong JP->WP translation fallbacks
Given Jetpack is merely a variant of WordPress right now, and given the way the string resources between `src/main/res` and `src/jetpack/res` are merged at compile time, this means that at runtime the Jetpack app will fallback to the WordPress translation of a string if a translation is not provided for that locale in `src/jetpack/res`, thus potentially containing wording about WordPress instead of Jetpack

Given that the strings in `src/jetpack/res` are especially provided for keys where an override IS needed for Jetpack (e.g. `app_name`), this could lead to inconsistencies.

To solve this, this commit makes a copy of the appropriate fallback of `src/jetpack/res/values*/strings.xml` (most oven `values` aka English) into the `src/jetpack/res/values-*` folders for which Jetpack otherwise doesn't support the locale of. That way, at runtime the Jetpack app will use that copied fallback and thus Jetpack-specific copies even if they are in English instead of using a translation in the user's locale but which would reference WordPress instead of Jetpack