Build: Enable building upon default config for detekt
Only those rules that have a different cofiguration will be now visible
within the 'detekt.yml' file. Those are the below:

- LongParameterList (complexity)
- ForbiddenComment (style)
- MagicNumber (style)

In addition to the above rules, the below generic configuration is added
to the 'detekt.yml' file as it also differs from the default:

- warningsAsErrors (config)
- android/autoCorrect (formatting)

Everything else matches to the default and as such is omitted from the
'detekt.yml' file.