Build: Update detekt to 1.16.0
With this update, some rules, like the below, which were disabled
by default, are now enabled by default instead:

- PrintStackTrace (exceptions)
- SwallowedException (exceptions)
- ImplicitDefaultLocale (potential-bugs)
- SerialVersionUIDInSerializableClass (style)
- UnusedPrivateMember (style)
- VarCouldBeVal (style)

As such, for all the above listed rules, Detekt has started to warn
about current violations based on these rules. As such those need to
be resolved/suppressed accordingly. This is going to be done in
subsequent commits.